+++ apps/api/src/tooling/tools/arxiv.py
@@ -1,0 +1,5 @@
+from langchain_community.utilities import ArxivAPIWrapper
+
+__all__ = ["ArxivAPIWrapper"]
+
+arxiv = ArxivAPIWrapper()
