+++ apps/api/src/tooling/tools/move_file.py
@@ -1,0 +1,5 @@
+from langchain_community.tools.file_management import MoveFileTool
+
+__all__ = ["MoveFileTool"]
+
+ID = "f57d47fd-5783-4aac-be34-17ba36bb6242"
