+++ apps/api/src/tooling/tools/read_file.py
@@ -1,0 +1,5 @@
+from langchain.tools.file_management.read import ReadFileTool
+
+__all__ = ["ReadFileTool"]
+
+ID = "ca16f5dd-c17f-4231-a3e6-4b6ddf2f3d67"
